+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `pycmIndexError` class
+- `pycmNotWorkingError` class
 - `MultiLabelCM` class
+- `classwise_cm` method
+- `samplewise_cm` method
 - `__mlcm_vector_handler__` function
 - `__mlcm_assign_classes__` function
 - `__mlcm_vectors_filter__` function