+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [4.0] - 2023-06-07
 ### Added
 - `pycmMultiLabelError` class
 - `MultiLabelCM` class
@@ -708,7 +709,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TPR
 - documents and `README.md`
 
-[Unreleased]: https://github.com/sepandhaghighi/pycm/compare/v3.9...dev
+[Unreleased]: https://github.com/sepandhaghighi/pycm/compare/v4.0...dev
+[4.0]: https://github.com/sepandhaghighi/pycm/compare/v3.9...v4.0
 [3.9]: https://github.com/sepandhaghighi/pycm/compare/v3.8...v3.9
 [3.8]: https://github.com/sepandhaghighi/pycm/compare/v3.7...v3.8
 [3.7]: https://github.com/sepandhaghighi/pycm/compare/v3.6...v3.7